move telemetry package to installer
