+++ cmd/installer-downloader/main.go
@@ -16,7 +16,7 @@ import (
 	"github.com/DataDog/datadog-agent/cmd/installer/user"
 	"github.com/DataDog/datadog-agent/pkg/fleet/installer/env"
 	"github.com/DataDog/datadog-agent/pkg/fleet/installer/oci"
-	"github.com/DataDog/datadog-agent/pkg/fleet/telemetry"
+	"github.com/DataDog/datadog-agent/pkg/fleet/installer/telemetry"
 )
 
 const (
