+++ pkg/fleet/installer/packages/app_armor.go
@@ -16,7 +16,7 @@ import (
 	"os/exec"
 	"strings"
 
-	"github.com/DataDog/datadog-agent/pkg/fleet/telemetry"
+	"github.com/DataDog/datadog-agent/pkg/fleet/installer/telemetry"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
