+++ pkg/fleet/installer/packages/file.go
@@ -15,7 +15,7 @@ import (
 	"os"
 	"syscall"
 
-	"github.com/DataDog/datadog-agent/pkg/fleet/telemetry"
+	"github.com/DataDog/datadog-agent/pkg/fleet/installer/telemetry"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
