+++ pkg/fleet/installer/telemetry/http_wrapper.go
