+++ pkg/fleet/installer/telemetry/span.go
