+++ pkg/fleet/installer/telemetry/telemetry.go
