+++ pkg/fleet/installer/telemetry/telemetry_test.go
