+++ pkg/fleet/installer/telemetry/tracer.go
